pyproject.toml
[tool.poetry]
name = "userclient"
version = "0.1.0"
description = ""
authors = ["linar"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.7.3"
pytest-aiohttp = "^0.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
